Write 201/202 entries, improve bands MIMO display, refactor
As per #6, this commit introduces the ability to specify an output mode
(`-m 201` or `-m 137`) so that different 00028874 files can be created.

By using mode 201, MIMO is taken into consideration and the parsing of a
combo string like `1A4A-1A4` will create two Band 1, Class A, 4x4 MIMO
downlink and a single band 1, Class A uplink.